# gmc_api_gateway/docs/swagger.yaml
basePath: /gmcapi/v2
definitions:
...
  model.CLOUD_DASHBOARD:
    properties:
      ClusterInfo: {}
      cpuTotal: {}
      cpuUsage: {}
      cpuUtil: {}
      diskTotal: {}
      diskUsage: {}
      diskUtil: {}
      memoryTotal: {}
      memoryUsage: {}
      memoryUtil: {}
      nodeInfo: {}
      nodeRunning: {}
      resourceCnt: {}
    type: object
...
  model.SERVICE_DASHBOARD:
    properties:
      podCpuTop5: {}
      podMemTop5: {}
      projectCnt:
        type: integer
      projectCpuTop5: {}
      projectList: {}
      projectMemTop5: {}
      resource: {}
      workspaceCnt:
        type: integer
    type: object
...
  model.TOTAL_DASHBOARD:
    properties:
      clusterCnt:
        type: integer
      clusterCpuTop5: {}
      clusterMemTop5: {}
      coreClusterCnt:
        type: integer
      credentialCnt:
        type: integer
      edgeClusterCnt:
        type: integer
      edgeInfo: {}
      podCpuTop5: {}
      podMemTop5: {}
      projectCnt:
        type: integer
      workspaceCnt:
        type: integer
    type: object
...
host: 101.79.1.138:8012
info:
  contact:
    name: GM-Center
    url: https://gedge-platform.github.io/gm-center/
  description: This is a Gedge GM-Center Swagger API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gedge GM-Center Swagger API
  version: "2.0"
paths:
...
  /cloudDashboard:
    get:
      consumes:
      - application/json
      description: get TotalDashboard info
      parameters:
      - description: name of the Cluster
        in: query
        name: cluster
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CLOUD_DASHBOARD'
      security:
      - Bearer: []
      summary: Show TotalDashboard
      tags:
      - Dashboard
...
  /serviceDashboard:
    get:
      consumes:
      - application/json
      description: get ServiceDashboard info
      parameters:
      - description: name of the User
        in: query
        name: user
        required: true
        type: string
      - description: name of the Workspace
        in: query
        name: workspace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SERVICE_DASHBOARD'
      security:
      - Bearer: []
      summary: Show ServiceDashboard
      tags:
      - Dashboard
...
  /totalDashboard:
    get:
      consumes:
      - application/json
      description: get TotalDashboard info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TOTAL_DASHBOARD'
      security:
      - Bearer: []
      summary: Show TotalDashboard
      tags:
      - Dashboard
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: '"Type \"Bearer \" and then your API Token"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"